+++ mayan/apps/ocr/handlers.py
@@ -11,7 +11,7 @@ from .settings import setting_auto_ocr
 logger = logging.getLogger(__name__)
 
 
-def handler_index_document(sender, **kwargs):
+def handler_index_document_version(sender, **kwargs):
     task_index_document.apply_async(
         kwargs=dict(document_id=kwargs['instance'].document.pk)
     )